requirements.txt
streamlit==1.45.1
pandas==2.2.3
openai==1.83.0
python-dotenv==1.1.0
rich==14.0.0
tqdm==4.67.1
openpyxl==3.1.5
geopy==2.4.1